+++ Bugsnag/Storage/BSGFileLocations.m
@@ -119,6 +119,7 @@ BSG_OBJC_DIRECT_MEMBERS
         _sessions = getAndCreateSubdir(root, @"sessions");
         _breadcrumbs = getAndCreateSubdir(root, @"breadcrumbs");
         _kscrashReports = getAndCreateSubdir(root, @"KSCrashReports");
+        _featureFlags = getAndCreateSubdir(root, @"featureFlags");
         _appHangEvent = [root stringByAppendingPathComponent:@"app_hang.json"];
         _flagHandledCrash = [root stringByAppendingPathComponent:@"bugsnag_handled_crash.txt"];
         _configuration = [root stringByAppendingPathComponent:@"config.json"];